+++ drivers/spi/spi_ambiq_bleif.c
@@ -14,7 +14,7 @@
 LOG_MODULE_REGISTER(spi_ambiq_bleif);
 
 #include <zephyr/drivers/spi.h>
-#include <zephyr/drivers/spi/rtio.h>
+#include "spi_rtio.h"
 #include <zephyr/drivers/pinctrl.h>
 #include <zephyr/kernel.h>
 #include <zephyr/sys/byteorder.h>
