+++ drivers/spi/spi_dw.c
@@ -31,7 +31,7 @@ LOG_MODULE_REGISTER(spi_dw);
 #endif
 
 #include <zephyr/drivers/spi.h>
-#include <zephyr/drivers/spi/rtio.h>
+#include "spi_rtio.h"
 #include <zephyr/irq.h>
 
 #include "spi_dw.h"
