+++ drivers/spi/spi_litex_common.h
@@ -6,7 +6,7 @@
 
 #include <zephyr/kernel.h>
 #include <zephyr/drivers/spi.h>
-#include <zephyr/drivers/spi/rtio.h>
+#include "spi_rtio.h"
 #include <zephyr/sys/byteorder.h>
 #include <zephyr/sys_clock.h>
