+++ drivers/spi/spi_mcux_flexio.c
@@ -9,7 +9,7 @@
 
 #include <errno.h>
 #include <zephyr/drivers/spi.h>
-#include <zephyr/drivers/spi/rtio.h>
+#include "spi_rtio.h"
 #include <zephyr/drivers/clock_control.h>
 #include <fsl_flexio_spi.h>
 #include <zephyr/logging/log.h>
