+++ drivers/spi/spi_npcx_spip.c
@@ -7,7 +7,7 @@
 #define DT_DRV_COMPAT nuvoton_npcx_spip
 
 #include <zephyr/drivers/spi.h>
-#include <zephyr/drivers/spi/rtio.h>
+#include "spi_rtio.h"
 #include <zephyr/drivers/clock_control.h>
 #include <zephyr/drivers/pinctrl.h>
 #include <zephyr/kernel.h>
