+++ drivers/spi/spi_nxp_lpspi/spi_nxp_lpspi_priv.h
@@ -8,12 +8,12 @@
 #define ZEPHYR_DRIVERS_SPI_SPI_NXP_LPSPI_PRIV_H_
 
 #include <zephyr/drivers/spi.h>
-#include <zephyr/drivers/spi/rtio.h>
 #include <zephyr/drivers/pinctrl.h>
 #include <zephyr/drivers/clock_control.h>
 #include <zephyr/sys/__assert.h>
 #include <zephyr/irq.h>
 
+#include "../spi_rtio.h"
 #include "../spi_context.h"
 
 #if CONFIG_NXP_LP_FLEXCOMM
