+++ drivers/spi/spi_renesas_rz_rspi.c
@@ -16,7 +16,7 @@
 #include "r_dmac_b.h"
 #endif
 #ifdef CONFIG_SPI_RTIO
-#include <zephyr/drivers/spi/rtio.h>
+#include "spi_rtio.h"
 #include <zephyr/rtio/rtio.h>
 #endif
 #include <zephyr/logging/log.h>
