+++ drivers/spi/spi_rv32m1_lpspi.c
@@ -10,7 +10,7 @@
 
 #include <errno.h>
 #include <zephyr/drivers/spi.h>
-#include <zephyr/drivers/spi/rtio.h>
+#include "spi_rtio.h"
 #include <zephyr/drivers/clock_control.h>
 #include <fsl_lpspi.h>
 #include <zephyr/drivers/pinctrl.h>
