+++ drivers/spi/spi_sam0.c
@@ -16,7 +16,7 @@ LOG_MODULE_REGISTER(spi_sam0);
 #include <errno.h>
 #include <zephyr/device.h>
 #include <zephyr/drivers/spi.h>
-#include <zephyr/drivers/spi/rtio.h>
+#include "spi_rtio.h"
 #include <zephyr/drivers/dma.h>
 #include <zephyr/drivers/pinctrl.h>
 #include <soc.h>
