+++ drivers/spi/spi_silabs_usart.c
@@ -14,7 +14,7 @@ LOG_MODULE_REGISTER(spi_silabs_usart);
 #include <zephyr/sys/sys_io.h>
 #include <zephyr/device.h>
 #include <zephyr/drivers/spi.h>
-#include <zephyr/drivers/spi/rtio.h>
+#include "spi_rtio.h"
 #include <soc.h>
 
 #include "em_cmu.h"
