+++ drivers/spi/spi_stm32.c
@@ -16,7 +16,7 @@ LOG_MODULE_REGISTER(spi_stm32);
 #include <zephyr/drivers/dma/dma_stm32.h>
 #include <zephyr/drivers/pinctrl.h>
 #include <zephyr/drivers/spi.h>
-#include <zephyr/drivers/spi/rtio.h>
+#include "spi_rtio.h"
 #include <zephyr/dt-bindings/memory-attr/memory-attr-arm.h>
 #include <zephyr/irq.h>
 #include <zephyr/kernel.h>
