+++ drivers/spi/spi_test.c
@@ -11,7 +11,7 @@
 
 #include <zephyr/kernel.h>
 #include <zephyr/drivers/spi.h>
-#include <zephyr/drivers/spi/rtio.h>
+#include "spi_rtio.h"
 
 #define DT_DRV_COMPAT vnd_spi
