+++ drivers/spi/spi_wch.c
@@ -13,7 +13,7 @@ LOG_MODULE_REGISTER(spi_wch);
 #include <errno.h>
 #include <zephyr/device.h>
 #include <zephyr/drivers/spi.h>
-#include <zephyr/drivers/spi/rtio.h>
+#include "spi_rtio.h"
 #include <zephyr/drivers/pinctrl.h>
 #include <zephyr/drivers/clock_control.h>
